// packages/design-system/src/styles/types.ts
import { Tokens } from "../tokens";

export type Theme = {
  borderStyles: Tokens["borderStyles"];
  borderWidths: Tokens["borderWidths"];
  breakPoints: Tokens["breakPoints"];
  durations: Tokens["durations"];
  easings: Tokens["easings"];
  fonts: Tokens["fonts"];
  fontSizes: Tokens["fontSizes"];
  fontWeights: Tokens["fontWeights"];
  lineHeights: Tokens["lineHeights"];
  letterSpacings: Tokens["letterSpacings"];
  opacity: Tokens["opacity"];
  radii: Tokens["radii"];
  shadows: Tokens["shadows"];
  spacing: Tokens["spacing"];
};